Implement path resolution and normalization utilities; refactor related code across filesystem and directory handling
