+++ lib/src/eval/runtime/security/permissions/filesystem.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'package:dart_eval/dart_eval_security.dart';
+import 'package:dart_eval/src/eval/utils/path_helper.dart';
 import 'package:path/path.dart' as p;
 
 /// A permission that allows access to read and write a file system resource.
@@ -40,55 +41,9 @@ class FilesystemPermission implements Permission {
     String normalized = p.normalize(absolutePath);
 
     // Try to resolve symlinks by walking up the directory tree
-    String resolved = _resolvePathRecursively(normalized);
+    String resolved = resolvePathRecursively(normalized);
 
     return resolved;
-  }
-
-  /// Recursively resolve symlinks by walking up the directory hierarchy
-  static String _resolvePathRecursively(String path) {
-    try {
-      // Try to resolve the full path directly first
-      if (File(path).existsSync()) {
-        return File(path).resolveSymbolicLinksSync();
-      } else if (Directory(path).existsSync()) {
-        return Directory(path).resolveSymbolicLinksSync();
-      }
-
-      // If the path doesn't exist, find the deepest existing parent and resolve it
-      String current = path;
-      List<String> nonExistentParts = [];
-
-      while (current != p.dirname(current)) {
-        // While not at root
-        // Add the current basename to non-existent parts FIRST
-        nonExistentParts.add(p.basename(current));
-
-        String parent = p.dirname(current);
-
-        if (Directory(parent).existsSync()) {
-          // Found an existing parent, resolve it and rebuild the path
-          String resolvedParent = Directory(parent).resolveSymbolicLinksSync();
-
-          // Rebuild the path with resolved parent + non-existent parts (in reverse order)
-          String result = resolvedParent;
-          for (String part in nonExistentParts.reversed) {
-            result = p.join(result, part);
-          }
-
-          return result;
-        }
-
-        // Move up to parent
-        current = parent;
-      }
-
-      // If we reach here, no parent could be resolved
-      return path;
-    } catch (e) {
-      // If any resolution fails, return the original path
-      return path;
-    }
   }
 
   /// Check if a given path is allowed by this permission
